Module Mh.Make

Metropolis-Hastings functor.

Parameters

module X : MH_parameters

Signature

val mcmc : verbosity:[ `Silent | `Progress | `Trace ] -> initial:X.t -> burn_in:int -> X.t Gen.t Gen.t

mcmc produces a sampler using Make_core_sampling_loop with the added guarantee that, provided the MH_parameters are consistent, the empirical measure will sample from the target measure specified by X.log_weight.

Be warned that the samples obtained from the chain are _not_ independent.The guarantee is that (if the proposal and weight are consistent) the empirical measure will converge in law to the target measure.