Module type Stats_intf.Fin_fun

Fin_fun is the module type describing the implementation of finite functions.

type t

t is the type of the domain of the finite function.

type r

r is the type of the range of the finite function. Typically, r will admit the structure of a ring.

module V : Basic_structures.Basic_intf.Free_module_std with type V.R.t = r and type V.Basis.t = t

V is the core of the implementation of the finite function, presented as a free R-module.

val total_weight : r

total_weight is the sum of the values of all points in the domain of the finite function.

val weightmap : V.t

weightmap is the value holding the finite function.