Module type Stats_intf.Fin_kernel

Fin_kernel is the module type packing the implementation of finite kernels.

type t

t is the type of the domain of the finite kernel.

type u

t is the type of the range of the finite kernel.

type r

r is the type of the measure. We expect r to admit the structure of a ring.

module V : Basic_structures.Basic_intf.Free_module_std with type V.R.t = r and type V.Basis.t = u

V is the core of the implementation of the finite functions, presented as a free R-module.

val kernel : t -> V.t

A kernel associates to each point in the domain a finitely supported function (typically interpreted as a finite measure).