Module type Stats_intf.Gen

Gen allows to manipulate generative probabilities (ie samplers).

type parameters = {

nsamples : int;

state : Stdlib.Random.State.t;

}

include Sampling_monad with type 'a Sampling_monad.res = parameters -> (module Basic_structures.Basic_intf.Std with type t = 'a) -> ('a, float) fin_mes

type !'a t = 'a gen

type !'a res = parameters -> (module Basic_structures.Basic_intf.Std with type t = 'a) -> ('a, float) fin_mes

val return : 'a -> 'a t

val bind : 'a t -> ('a -> 'b t) -> 'b t

val map : 'a t -> ('a -> 'b) -> 'b t

val run : 'a t -> 'a res

module Infix : sig ... end

val float : float -> float t

float bound samples uniformly in 0; bound

val int : int -> int t

int bound samples uniformly in 0; bound

val bool : bool t

bool samples a boolean uniformly

val uniform : 'a array -> 'a t

uniform elts samples an element of the elts array by sampling an index uniformly.

raises Invalid_argument

if elts has length 0

val bernouilli : float -> bool t

bernouilli alpha samples true with probability alpha.

raises Invalid_argument

if alpha is not in the 0;1 interval.

val geometric : float -> int t

geometric p samples a nonnegative integer according to the geometric law of parameter p.

val subsample : n:int -> 'a t -> 'a t

subsample ~n gen samples one out of n samples from gen.

val of_empirical : 'a emp -> 'a t

of_empirical emp samples from emp matching exactly the empirical frequencies.

val exponential : rate:float -> float t

Exponential distribution via inverse CDF.

val box_muller : mean:float -> std:float -> (float * float) t

Gaussian distribution via Box-Muller transform. Returns a pair of _independent_ gaussian variates with prescribed mean and standard deviation.

val gaussian : mean:float -> std:float -> float t

Gaussian distribution (wrapper over box-muller transform).

val range : range -> float t

Samples uniformly in the given range.

raises Invalid_argument

if range is empty.

val categorical : ('a * float) list -> 'a t

Categorical distribution. Total mass need not be one. Does not aggregate mass of equal elements.

raises Invalid_argument

if some weights are negative or if the total mass is zero.

val of_fin_mes : ('a, float) fin_mes -> 'a t

Constructs a sampler from a finite measure through categorical. NB: this is mostly useful for sampling repeatedly from the distribution.

module Rational : sig ... end