Module Stats__Fin

Finite measures implemented as finitely supported functions.

Finitely supported distributions.

Generic constructors.

val measure : (module Basic_structures.Basic_intf.Reals with type t = 'r) -> (module Basic_structures.Basic_intf.Free_module_std with type Basis.t = 't and type R.t = 'r) -> ('t * 'r) list -> ('t, 'r) Stats.Stats_intf.fin_mes

measure reals vecs data returns the finite measure associated to data. data is a list of weighted elements, with weights admitting the structure specifiedby reals. An implementation of free modules over reals must provided as vecs.

raises Invalid_argument

if a weight is negative.

val probability : (module Basic_structures.Basic_intf.Reals with type t = 'r) -> (module Basic_structures.Basic_intf.Free_module_std with type Basis.t = 't and type R.t = 'r) -> ('t * 'r) list -> ('t, 'r) Stats.Stats_intf.fin_prb

probability reals vecs data packs data as a fin_prb.

raises Invalid_argument

if the sum of the weights is not equal to one or if a weight is negative.

val as_measure : ('a, 'b) Stats.Stats_intf.fin_prb -> ('a, 'b) Stats.Stats_intf.fin_mes

Forgetful map from the type of finite probabilities to the type of measures.

Functions specific to the underlying field.

module Float : sig ... end

float-valued finitely supported distributions.

module Rational : sig ... end

Q-valued finitely supported distributions.