Module Stats__Graph.Dist

Dist handles distances between vertices. The distance between two vertices is the length of a shortest path between those vertices.

type t =

| Inf

| Fin of int

val zero : t

The distance between two vertices is zero iff they are equal.

val one : t

The distance between two vertices is one iff they are adjacent.

val infty : t

The distance between two vertices is infty iff they are disconnected.

val (+) : t -> t -> t

Adding distances.

val (>) : t -> t -> bool

Comparing distances.

val max : t -> t -> t

Computing the max of two distances.