Module Stats__Stats_intf

Type signatures.

type range = {

min : float;

max : float;

}

The type of (inclusive) ranges.

type 'a emp = 'a array

'a emp is the type of empirical measures

type 'a gen = Stdlib.Random.State.t -> 'a

'a gen is the type of samplers

module type Fin_fun = sig ... end

Fin_fun is the module type describing the implementation of finite functions.

module type Fin_kernel = sig ... end

Fin_kernel is the module type packing the implementation of finite kernels.

type ('a, 'r) fin_fun = (module Fin_fun with type r = 'r and type t = 'a)

('a, 'r) fin_fun is the type of finite functions from 'a to 'r.

type ('a, 'r) fin_mes =

| M of ('a, 'r) fin_fun

fin_mes is the type of finitely supported measures.

type ('a, 'r) fin_prb =

| P of ('a, 'r) fin_fun

fin_mes is the type of finitely supported probability measures.

type ('a, 'b, 'r) kernel = (module Fin_kernel with type r = 'r and type t = 'a and type u = 'b)

('a, 'b,'r) is the type of finite kernels from 'a to 'b with measure valued in 'r.

module type Sampling_monad = Basic_structures.Basic_intf.Monad with type 'a Sampling_monad.t = 'a gen

Sampling_monad is the module type of a sampling-based monad

module type Gen = sig ... end

Gen allows to manipulate generative probabilities (ie samplers).

module type Fin_dist = sig ... end

module type Graph = sig ... end

We use an OCamlgraph-compatible module type to describe undirected graphs. We assume that all graphs are undirected and simple.