Module Mh.Make_core_sampling_loop

Generic sampling loop.

Parameters

X : Basic_structures.Basic_intf.Pp

Signature

val mcmc : verbosity:[ `Silent | `Progress | `Trace ] -> initial:X.t -> burn_in:int -> (X.t -> X.t Stats__.Gen.t) -> X.t Stats__.Gen.t Stats__.Gen.t

The sampler function produces a generative probability. More precisely: each sample from mcmc ~verbosity ~initial ~burn_in is a distinct Markov chain, which samples can in-turn be obtained by calling iteratively the closure returned by mcmc. Concretely, the outer `Stats.gen` corresponds to sampling the burn-in while the inner corresponds to sampling from the actual chain.